+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+
 ## [0.3.8-alpha.1] 2022-11-16
 
 ### Added
@@ -26,6 +27,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [System] Improve the feedback function, you can submit feedback in the bottom right corner of WebUI.
+
+### Fixed
+- [System] Fix CPU Temp for other platforms ([#661](https://github.com/IceWhaleTech/CasaOS/issues/661))
 
 ## [0.3.7.1] 2022-11-04
 
@@ -47,7 +51,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [Apps] App list update mechanism improved, now you can see the latest apps in App Store immediately.
 - [Storage] Fixed a lot of known issues
-
 
 ### Added
 - [Storage] Disk merge (Beta), you can merge multiple disks into a single storage space (currently you need to enable this feature from the command line)